Rid off IE stub classes
