+++ src/main/java/me/jellysquid/mods/sodium/client/gui/options/Option.java
@@ -1,13 +1,20 @@
 package me.jellysquid.mods.sodium.client.gui.options;
 
+import me.jellysquid.mods.sodium.client.SodiumClientMod;
 import me.jellysquid.mods.sodium.client.gui.options.control.Control;
 import me.jellysquid.mods.sodium.client.gui.options.storage.OptionStorage;
 import net.minecraft.network.chat.Component;
-import org.embeddedt.embeddium.gui.IEOption;
+import net.minecraft.resources.ResourceLocation;
 
 import java.util.Collection;
 
-public interface Option<T> extends IEOption<T> {
+public interface Option<T>{
+    ResourceLocation DEFAULT_ID = new ResourceLocation(SodiumClientMod.MODID, "empty");
+
+    default ResourceLocation getId() {
+        return DEFAULT_ID;
+    }
+
     Component getName();
 
     Component getTooltip();
